Program service, vice lektoru

+++ app/AdminModule/ProgramModule/components/ProgramCategoriesGridControl.php
@@ -6,12 +6,10 @@ namespace App\AdminModule\ProgramModule\Components;
 
 use App\Model\ACL\Role;
 use App\Model\ACL\RoleRepository;
-use App\Model\Program\Category;
 use App\Model\Program\CategoryRepository;
 use App\Model\Program\ProgramRepository;
 use App\Model\User\UserRepository;
 use App\Services\ProgramService;
-use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\OptimisticLockException;
 use Doctrine\ORM\ORMException;
 use Kdyby\Translation\Translator;
@@ -137,15 +135,9 @@ class ProgramCategoriesGridControl extends Control
      */
     public function add(\stdClass $values) : void
     {
-        $category = new Category();
+        $this->programService->createCategory($values['name'], $this->roleRepository->findRolesByIds($values['registerableRoles']));
 
-        $category->setName($values['name']);
-        $category->setRegisterableRoles($this->roleRepository->findRolesByIds($values['registerableRoles']));
-
-        $this->categoryRepository->save($category);
-
-        $p = $this->getPresenter();
-        $p->flashMessage('admin.program.categories_saved', 'success');
+        $this->getPresenter()->flashMessage('admin.program.categories_saved', 'success');
 
         $this->redirect('this');
     }
@@ -159,16 +151,7 @@ class ProgramCategoriesGridControl extends Control
     {
         $category = $this->categoryRepository->findById($id);
 
-        $this->categoryRepository->getEntityManager()->transactional(function ($em) use ($category, $values) : void {
-            $category->setName($values['name']);
-            $category->setRegisterableRoles($this->roleRepository->findRolesByIds($values['registerableRoles']));
-
-            $this->categoryRepository->save($category);
-
-            $this->programService->updateUsersPrograms(new ArrayCollection($this->userRepository->findAll()));
-
-            $this->categoryRepository->save($category);
-        });
+        $this->programService->updateCategory($category, $values['name'], $this->roleRepository->findRolesByIds($values['registerableRoles']));
 
         $this->getPresenter()->flashMessage('admin.program.categories_saved', 'success');
         $this->redirect('this');
@@ -183,13 +166,13 @@ class ProgramCategoriesGridControl extends Control
     {
         $category = $this->categoryRepository->findById($id);
 
-        $this->categoryRepository->getEntityManager()->transactional(function ($em) use ($category) : void {
+        if ($category->getBlocks()->isEmpty()) {
             $this->categoryRepository->remove($category);
+            $this->getPresenter()->flashMessage('admin.program.categories_deleted', 'success');
+        } else {
+            $this->getPresenter()->flashMessage('admin.program.categories_deleted_error', 'danger');
+        }
 
-            $this->programService->updateUsersPrograms(new ArrayCollection($this->userRepository->findAll()));
-        });
-
-        $this->getPresenter()->flashMessage('admin.program.categories_deleted', 'success');
         $this->redirect('this');
     }
 }
